# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Your API Documentation",
    "version": "1.0.0",
    "description": "API documentation for your Express.js app"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "http://localhost:50000"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "login": {
        "type": "object",
        "properties": {
          "username": {
            "$ref": "#/components/schemas/username"
          },
          "password": {
            "$ref": "#/components/schemas/password"
          }
        }
      },
      "signIn": {
        "type": "object",
        "properties": {
          "username": {
            "$ref": "#/components/schemas/username"
          },
          "mail": {
            "$ref": "#/components/schemas/mail"
          },
          "password": {
            "$ref": "#/components/schemas/password"
          }
        }
      },
...
      "getPost": {
        "type": "object",
        "properties": {
          "id_post": {
            "$ref": "#/components/schemas/id"
          },
          "title": {
            "$ref": "#/components/schemas/title"
          },
          "content": {
            "$ref": "#/components/schemas/content"
          },
          "publish_date": {
            "$ref": "#/components/schemas/publish_date"
          },
          "publisher": {
            "$ref": "#/components/schemas/username"
          },
          "plays": {
            "$ref": "#/components/schemas/amount"
          },
          "signals": {
            "$ref": "#/components/schemas/amount"
          }
        }
      },
      "newPost": {
        "type": "object",
        "properties": {
          "title": {
            "$ref": "#/components/schemas/title"
          },
          "content": {
            "$ref": "#/components/schemas/content"
          },
          "qsetArray": {
            "$ref": "#/components/schemas/postQsetArray"
          }
        }
      },
...
      "postQsetArray": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/qsetNoID"
        }
      },
...
      "vgd": {
        "type": "object",
        "properties": {
          "id_vg": {
            "$ref": "#/components/schemas/id"
          },
          "name": {
            "$ref": "#/components/schemas/name"
          },
          "image_link": {
            "$ref": "#/components/schemas/link"
          },
          "release_date": {
            "$ref": "#/components/schemas/release_date"
          },
          "description": {
            "$ref": "#/components/schemas/description"
          }
        }
      },
      "vgdNoID": {
        "type": "object",
        "properties": {
          "name": {
            "$ref": "#/components/schemas/name"
          },
          "image_link": {
            "$ref": "#/components/schemas/link"
          },
          "release_date": {
            "$ref": "#/components/schemas/release_date"
          },
          "description": {
            "$ref": "#/components/schemas/description"
          }
        }
      },
...
      "qsetNoID": {
        "type": "object",
        "properties": {
          "prompt": {
            "$ref": "#/components/schemas/prompt"
          },
          "ressource_type": {
            "$ref": "#/components/schemas/ressource_type"
          },
          "ressource_link": {
            "$ref": "#/components/schemas/link"
          },
          "questionArray": {
            "$ref": "#/components/schemas/questionArrayNoID"
          }
        }
      },
...
      "questionArrayNoID": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/questionNoID"
        }
      },
...
      "questionNoID": {
        "type": "object",
        "properties": {
          "is_correct": {
            "$ref": "#/components/schemas/is_correct"
          },
          "statement": {
            "$ref": "#/components/schemas/statement"
          }
        }
      },
      "q": {
        "type": "string",
        "example": "Query%20Example",
        "description": "The query is url encoded"
      },
      "gs": {
        "type": "string",
        "example": "0%201%208",
        "description": "The list is separated by space %20"
      },
      "username": {
        "type": "string",
        "example": "myusername"
      },
      "mail": {
        "type": "string",
        "example": "mail@example.com"
      },
      "password": {
        "type": "string",
        "format": "password",
        "example": "clearpassword"
      },
      "token": {
        "type": "string",
        "example": "jwt token like (TODO)"
      },
      "id": {
        "type": "integer",
        "example": 1
      },
      "title": {
        "type": "string",
        "example": "Hello World !"
      },
      "content": {
        "type": "string",
        "example": "My name is Nico."
      },
      "is_correct": {
        "type": "boolean",
        "example": true
      },
      "statement": {
        "type": "string",
        "example": "A example of an affirmation."
      },
      "prompt": {
        "type": "string",
        "example": "What is the correct answer ?"
      },
      "ressource_type": {
        "type": "string",
        "example": "image"
      },
      "link": {
        "type": "string",
        "example": "https://cdn.example.com/image.jpg"
      },
      "publish_date": {
        "type": "string",
        "format": "date-time",
        "example": "2017-07-21T17:32:28Z"
      },
      "amount": {
        "type": "integer",
        "example": 169
      },
      "played": {
        "type": "boolean",
        "example": true
      },
      "signaled": {
        "type": "boolean",
        "example": true
      },
      "name": {
        "type": "string",
        "example": "Zelda"
      },
      "release_date": {
        "type": "string",
        "format": "date",
        "example": "2017-07-21"
      },
      "description": {
        "type": "string",
        "example": "A amazing description of a videogame"
      },
...
    },
    "responses": {
      "401": {
        "description": "Unauthorized, must provide an authentification bearer jwt"
      },
      "403": {
        "description": "Although your token is valid, you don't the rights to do so"
      }
    }
  },
  "paths": {
    "/auth/login": {
      "post": {
        "security": [],
        "description": "Login with body credentials",
        "tags": [
          "Authenfication"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/login"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful login !",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "$ref": "#/components/schemas/token"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Failed to login with these creds."
          },
          "404": {
            "description": "User not found !"
          }
        }
      }
    },
    "/auth/sign-in": {
      "post": {
        "security": [],
        "description": "Sign-in to the website",
        "tags": [
          "Authenfication"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/signIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "You account have been created!"
          },
          "400": {
            "description": "Bad information provided"
          }
        }
      }
    },
    "/post": {
      "get": {
        "description": "Query some posts",
        "tags": [
          "Post"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "gs",
            "description": "The gamescope is selection of the game you want to search for. The list is separated by dash -",
            "schema": {
              "$ref": "#/components/schemas/gs"
            }
          },
          {
            "in": "query",
            "name": "q",
            "description": "The query (q) parameter is based text you want to search for in the title.",
            "schema": {
              "$ref": "#/components/schemas/q"
            }
          },
          {
            "in": "query",
            "name": "u",
            "description": "The user parameter is to select post for only one username.",
            "schema": {
              "$ref": "#/components/schemas/username"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the result from db",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/getPost"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad information provided"
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      },
      "post": {
        "description": "Post a new post",
        "tags": [
          "Post"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/newPost"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "New post created !"
          },
          "400": {
            "description": "Bad information provided"
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      }
    },
...
    "/post/{postId}/play": {
      "parameters": [
        {
          "in": "path",
          "name": "postId",
          "schema": {
            "$ref": "#/components/schemas/id"
          },
          "required": true,
          "description": "The post Id to manage"
        }
      ],
      "get": {
        "description": "Get amount of plays (likes) of a specific post and if I played",
        "tags": [
          "Post"
        ],
        "responses": {
          "200": {
            "description": "Return the result from db",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "amount": {
                      "$ref": "#/components/schemas/amount"
                    },
                    "played": {
                      "$ref": "#/components/schemas/played"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "description": "Post not found"
          }
        }
      },
      "post": {
        "description": "User play (like) of a specific post",
        "tags": [
          "Post"
        ],
        "responses": {
          "201": {
            "description": "New play created !"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "description": "Post not found"
          }
        }
      },
      "delete": {
        "description": "User retrieve play (like) of a specific post",
        "tags": [
          "Post"
        ],
        "responses": {
          "200": {
            "description": "Deleted play !"
          },
          "400": {
            "description": "Cannot delete your (none) play - please play first"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "description": "Post not found"
          }
        }
      }
    },
    "/post/{postId}/signal": {
      "parameters": [
        {
          "in": "path",
          "name": "postId",
          "schema": {
            "$ref": "#/components/schemas/id"
          },
          "required": true,
          "description": "The post Id to manage"
        }
      ],
      "get": {
        "description": "Get amount of signals of a specific post",
        "tags": [
          "Post"
        ],
        "responses": {
          "200": {
            "description": "Return the result from db",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "amount": {
                      "$ref": "#/components/schemas/amount"
                    },
                    "signaled": {
                      "$ref": "#/components/schemas/signaled"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "description": "Post not found"
          }
        }
      },
      "post": {
        "description": "User signal a specific post",
        "tags": [
          "Post"
        ],
        "responses": {
          "201": {
            "description": "New signal created !"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "description": "Post not found"
          }
        }
      }
    },
    "/vgd": {
      "get": {
        "description": "Get all video games available",
        "tags": [
          "VideoGame Discovery"
        ],
        "responses": {
          "200": {
            "description": "Fetch all videogames",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/vgd"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      },
      "post": {
        "description": "Post a new video game",
        "tags": [
          "VideoGame Discovery"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/vgdNoID"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Modify a video game"
          },
          "400": {
            "description": "Bad informations provided"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          }
        }
      },
      "put": {
        "description": "Modify a video game",
        "tags": [
          "VideoGame Discovery"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/vgd"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create a new video game"
          },
          "400": {
            "description": "Bad informations provided"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}